requirements.txt
streamlit
phidata
openai
google-generativeai
duckduckgo-search